pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ces-model"
version = "0.1.0"
description = "CES substitution model: Brent crude oil price impact on renewable energy investment"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "research-team"}]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "SALib>=1.4",
    "arch>=5.0",
    "matplotlib>=3.7",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
volatility-ml = [
    "xgboost>=1.7",
    "shap>=0.42",
]
dev = [
    "ruff",
    "pytest>=7.0",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ces_model"]

[tool.ruff]
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.9",
]
